feat(transaction): sécurise les autorisations et corrige les tests refund
- Ajout de `authorize('refund')` explicite dans le contrôleur TransactionController
- Correction des tests : ajout de `actingAs()` pour garantir le bon utilisateur
- Meilleure couverture des cas 403 vs 422 dans les refund non autorisés
- Ajout d’un test pour `create` bloqué si utilisateur non vérifié (policy)
- Ajustement du `return` dans store() avec type `TransactionResource::make(...)`

+++ database/migrations/0001_01_01_000010_create_transactions_table.php
@@ -26,7 +26,7 @@ return new class extends Migration
                 ->nullOnDelete();
 
             // 💰 Montant + devise
-            $table->decimal('amount', 10, 2);
+            $table->float('amount');
             $table->string('currency', 10)
                 ->default(CurrencyEnum::EUR->value);
